FIX: adjust in MaterialDetailModel quantity type to BigDecimal

+++ ProdOrderManager/src/main/java/com/AC/ProdOrderManager/dtos/material/RegisterMaterialDetailRequestDTO.java
@@ -2,5 +2,5 @@ package com.AC.ProdOrderManager.dtos.material;
 
 import java.math.BigDecimal;
 
-public record RegisterMaterialDetailRequestDTO(String baseMaterialName, Integer quantity, BigDecimal totalAmount, BigDecimal paidAmount) {
+public record RegisterMaterialDetailRequestDTO(String baseMaterialName, BigDecimal quantity, BigDecimal totalAmount, BigDecimal paidAmount) {
 }
